Refactor public API: typed wrappers, Swift-native Int, enum extensions, boxed types
Replace all OpaquePointer with typed Swift wrappers (TabPage, ViewStack,
Carousel, NavigationPage, Banner, Dialog, etc.) and add init(borrowing:)
to all subclasses that override init(raw:). Convert all Int32/UInt32 in
public API to Int (~110 places). Add Swift enum extensions for 17 C enum
types enabling dot-syntax (.vertical, .center, .crossfade). Create
SpringParams and BreakpointCondition boxed-type wrappers. Add convenience
methods on Widget (tooltipText, setSizeRequest, root, opacity) and
ListBox (unselectAll, selectAll). Update DemoApp to use new APIs.

+++ Sources/Adwaita/ApplicationWindow.swift
@@ -44,30 +44,30 @@ public final class ApplicationWindow: Widget {
     }
 
     /// The default width of the window.
-    public var defaultWidth: Int32 {
+    public var defaultWidth: Int {
         get {
             var w: Int32 = 0
             gtk_window_get_default_size(castedPointer(), &w, nil)
-            return w
+            return Int(w)
         }
         set {
             var h: Int32 = 0
             gtk_window_get_default_size(castedPointer(), nil, &h)
-            gtk_window_set_default_size(castedPointer(), newValue, h)
+            gtk_window_set_default_size(castedPointer(), Int32(newValue), h)
         }
     }
 
     /// The default height of the window.
-    public var defaultHeight: Int32 {
+    public var defaultHeight: Int {
         get {
             var h: Int32 = 0
             gtk_window_get_default_size(castedPointer(), nil, &h)
-            return h
+            return Int(h)
         }
         set {
             var w: Int32 = 0
             gtk_window_get_default_size(castedPointer(), &w, nil)
-            gtk_window_set_default_size(castedPointer(), w, newValue)
+            gtk_window_set_default_size(castedPointer(), w, Int32(newValue))
         }
     }
